OBS::WebSocket::Connection added
OBS::WebSocket::Connection added.
Connection constructor accepts uri, password as keyword arguments (supports loading conn info from config)

default port for TCPSocket changed to 4455.

SocketWrapper class moved into Connection class.

websocket-driver example updated to reflect changes.

'websocket-driver' added as dependency in gemspec file
