+++ .gitignore
@@ -11,3 +11,6 @@
 .rspec_status
 
 *.bak
+
+examples/from-toml/
+.editorconfig
