+++ examples/websocket-driver/Gemfile
@@ -1,4 +1,5 @@
-source 'https://rubygems.org'
+# frozen_string_literal: true
+
+source "https://rubygems.org"
 
 gem 'obs-websocket', path: '../..'
-gem 'websocket-driver'
