+++ obs-websocket.gemspec
@@ -29,4 +29,5 @@ Gem::Specification.new do |spec|
   spec.require_paths = ['lib']
 
   spec.add_dependency 'concurrent-ruby', '~> 1.1'
+  spec.add_dependency 'websocket-driver', '~> 0.7.5'
 end
